+++ src/components/Profile.tsx
@@ -8,7 +8,7 @@ import {
   Title,
 } from "@mantine/core";
 import { IconBrandTwitter } from "@tabler/icons-react";
-import { useFirebaseAuth } from "lib/auth/auth";
+import { useFirebaseAuth } from "lib/auth/firebaseAuth";
 import { useMediaQuery } from "lib/mantine/useMediaQuery";
 
 const Profile = () => {
@@ -71,18 +71,18 @@ const Profile = () => {
             </Text>
           </div>
           <div className="mx-auto flex w-1/2 items-center justify-center">
-            {/* {twitterUsername && (
+            {twitter_name && (
               <>
                 <ActionIcon
                   component="a"
-                  href={`https://twitter.com/${twitterUsername}`}
+                  href={`https://twitter.com/${twitter_name}`}
                   target="_blank"
                 >
                   <IconBrandTwitter size={18} />
                 </ActionIcon>
-                <Text size="sm">@{twitterUsername}</Text>
+                <Text size="sm">@{twitter_name}</Text>
               </>
-            )} */}
+            )}
           </div>
 
           <Button variant="default" fullWidth mt="md">
