+++ src/features/auth/components/LoginModal.tsx
@@ -1,11 +1,6 @@
 import { Group, Image, Modal, Stack, Text } from "@mantine/core";
 import GoogleButton from "components/Button/GoogleButton";
-import {
-  getAdditionalUserInfo,
-  getAuth,
-  GoogleAuthProvider,
-  signInWithPopup,
-} from "firebase/auth";
+import { signInWithGoogle } from "lib/auth/auth";
 import { FC } from "react";
 import { useNavigate } from "react-router-dom";
 
@@ -18,29 +13,6 @@ type LoginModalProps = {
 
 const LoginModal: FC<LoginModalProps> = ({ opened, setOpened }) => {
   const navigate = useNavigate();
-  const signInwithGoogle = () => {
-    const auth = getAuth();
-    const provider = new GoogleAuthProvider();
-    signInWithPopup(auth, provider)
-      .then((result) => {
-        setOpened(false);
-        if (getAdditionalUserInfo(result)?.isNewUser) {
-          navigate("/acquisition");
-          return;
-        }
-        navigate("/");
-      })
-      .catch((error) => {
-        setOpened(false);
-        if (error.code === "auth/account-exists-with-different-credential") {
-          alert(
-            `${error.customData.email}は他のSNSと連携した既存ユーザーが登録済みです。既存ユーザーでログイン後、こちらのSNSとの連携が可能です`
-          );
-          return;
-        }
-        alert(`ログイン/新規登録に失敗しました。\n${error.message}`);
-      });
-  };
 
   return (
     <Modal opened={opened} onClose={() => setOpened(false)} centered size="sm">
@@ -50,7 +22,10 @@ const LoginModal: FC<LoginModalProps> = ({ opened, setOpened }) => {
           GameArchiveは、あなたのゲーム進行状況をシンプルに管理するアプリです。各ゲームの状態を一元管理し、公開設定を選んだユーザーは他のプレイヤーとゲームプレイ状況を共有できます。
         </Text>
         <Group grow mb="md" mt="md">
-          <GoogleButton title="Login With Google" onClick={signInwithGoogle} />
+          <GoogleButton
+            title="Login With Google"
+            onClick={() => signInWithGoogle(setOpened, navigate)}
+          />
         </Group>
         <Text size="sm" color="dimmed" className="px-5 text-center">
           利用規約、プライバシーポリシーに同意したうえでログインしてください。
